requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
flake8==7.1.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.1.0
propcache==0.2.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.0
